refactor: centralize fallback relay handling and replace inline logic with reusable methods
- Added `fetchAccountRelayListEvent` helper for kind 10002 relay list retrieval.
- Introduced `getFallbackRelays` method in settings store for streamlined relay access.
- Replaced redundant fallback relay retrieval logic across services with new abstractions.
- Improved maintainability and reduced code duplication with reusable patterns.
